[#22] chore : Home -> home 폴더 소문자로 변경 (#23) (#24)
