--- src/screens/Home/Home.style.ts
@@ -1,10 +1,0 @@
-import NaverMapView from 'react-native-nmap'
-import { SafeAreaView } from 'react-native-safe-area-context'
-import styled from 'styled-components/native'
-
-export const HomeLayout = styled(SafeAreaView)`
-  flex: 1;
-`
-export const MapView = styled(NaverMapView)`
-  flex: 1;
-`
