--- src/screens/Home/Home.tsx
@@ -1,12 +1,0 @@
-import NaverMap from './components/NaverMap'
-import * as S from '@/screens/Home/Home.style'
-
-const Home = () => {
-  return (
-    <S.HomeLayout>
-      <NaverMap />
-    </S.HomeLayout>
-  )
-}
-
-export default Home
